Solution to Manage Book Library

+++ Manage Book Library@@67877d61dc107486ad5c274c/code.c
@@ -10,8 +10,8 @@ int main() {
         scanf("%s",&author);
         float price;
         scanf("%0.2f",&price);
-        int threshold;
-        scanf("%d",&threshold);
+        float threshold;
+        scanf("%0.2f",&threshold);
         if (price>threshold){
             printf("Books above price %0.2f:\n",threshold);
             printf("Title: %s, Author: %s, Price:%0.2f\n",book,author,price);
